pyproject.toml
[project]
name = "piscine-datascience"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pyside6>=6.9.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
